Extract precommit_setup module from session_start
Move PrecommitNotInstalled, PrecommitInstallingHooks, PrecommitSetup,
and the install function into tools/claude_hooks/precommit_setup.py,
following the 1-module-1-setup-component pattern used by bazelisk_setup,
nix_setup, podman_service, and proxy_setup.
