+++ tools/claude_hooks/BUILD.bazel
@@ -128,6 +128,14 @@ py_library(
 )
 
 py_library(
+    name = "precommit_setup",
+    srcs = ["precommit_setup.py"],
+    imports = ["../.."],
+    visibility = ["//visibility:public"],
+    deps = ["@pypi//pydantic"],
+)
+
+py_library(
     name = "podman_service",
     srcs = ["podman_service.py"],
     imports = ["../.."],
@@ -176,6 +184,7 @@ py_library(
         ":errors",
         ":nix_setup",
         ":podman_service",
+        ":precommit_setup",
         ":proxy_credentials",
         ":proxy_setup",
         ":settings",
@@ -228,6 +237,7 @@ py_package(
         ":managed_files",
         ":nix_setup",
         ":podman_service",
+        ":precommit_setup",
         ":proxy_credentials",
         ":proxy_setup",
         ":proxy_vars",
