+++ tools/claude_hooks/session_start.py
@@ -26,7 +26,7 @@ from pydantic import BaseModel
 from fmt_util.fmt_util import format_limited_list
 from tools import env_utils
 from tools.build_info import BUILD_COMMIT
-from tools.claude_hooks import bazelisk_setup, env_file, nix_setup, podman_service, proxy_setup
+from tools.claude_hooks import bazelisk_setup, env_file, nix_setup, podman_service, precommit_setup, proxy_setup
 from tools.claude_hooks.debug import log_entrypoint_debug
 from tools.claude_hooks.errors import DirenvError, SkipError
 from tools.claude_hooks.settings import HookSettings
@@ -204,7 +204,7 @@ def emit_session_context(
     log_file: Path,
     proxy: proxy_setup.ProxySetup,
     podman: podman_service.PodmanSetup | None,
-    precommit: PrecommitSetup | None,
+    precommit: precommit_setup.PrecommitSetup | None,
 ) -> None:
     """Emit compact context summary for Claude Code transcript.
 
@@ -221,94 +221,13 @@ def emit_session_context(
         proxy=proxy,
         podman=podman,
         precommit=precommit,
-        PrecommitInstallingHooks=PrecommitInstallingHooks,
+        PrecommitInstallingHooks=precommit_setup.PrecommitInstallingHooks,
         log_entries=collector.buffer,
         has_github_token=bool(os.environ.get("DUCKTAPE_CI_READ_GITHUB_TOKEN")),
         log_file=log_file,
     )
     print(result.rstrip("\n"))
     sys.stdout.flush()
-
-
-class PrecommitNotInstalled(BaseModel):
-    """pre-commit hook was not installed (binary missing, install failed, etc.)."""
-
-    kind: Literal["not_installed"] = "not_installed"
-
-
-class PrecommitInstallingHooks(BaseModel):
-    """pre-commit hook installed and background `install-hooks` is running."""
-
-    kind: Literal["installing_hooks"] = "installing_hooks"
-    pid: int
-
-
-PrecommitSetup = PrecommitNotInstalled | PrecommitInstallingHooks
-
-
-def install_git_precommit_hook(project_dir: Path) -> PrecommitSetup:
-    """Install git pre-commit hook using the pre-commit CLI.
-
-    If `pre-commit` is not on PATH, installs it via pip first.
-    pre-commit is not pre-installed in Claude Code on the web's container.
-    pre-commit itself handles missing .git, missing config, and idempotent installs.
-
-    After installing the hook, launches `pre-commit install-hooks` in the background
-    to eagerly pre-install hook environments so the first commit doesn't pay that cost.
-    """
-    precommit = shutil.which("pre-commit")
-    if not precommit:
-        logger.info("pre-commit not found on PATH, installing via pip...")
-        try:
-            subprocess.run(["pip", "install", "pre-commit"], check=True, capture_output=True, text=True, timeout=120)
-            logger.info("Installed pre-commit via pip")
-        except (subprocess.CalledProcessError, subprocess.TimeoutExpired, FileNotFoundError) as e:
-            logger.warning("Failed to install pre-commit via pip: %s", e)
-            return PrecommitNotInstalled()
-        precommit = shutil.which("pre-commit")
-        if not precommit:
-            logger.warning("pre-commit still not found on PATH after pip install")
-            return PrecommitNotInstalled()
-
-    try:
-        version = subprocess.run([precommit, "--version"], capture_output=True, text=True, check=True, timeout=5)
-        logger.info("pre-commit %s", version.stdout.strip())
-    except (subprocess.CalledProcessError, subprocess.TimeoutExpired):
-        pass
-
-    hook_installed = False
-    try:
-        install_result = subprocess.run(
-            [precommit, "install"], check=False, cwd=project_dir, capture_output=True, text=True, timeout=30
-        )
-        if install_result.returncode == 0:
-            logger.info("Installed git pre-commit hook")
-            hook_installed = True
-        else:
-            logger.warning("pre-commit install failed: %s", install_result.stderr)
-    except subprocess.TimeoutExpired:
-        logger.warning("pre-commit install timed out")
-
-    if not hook_installed:
-        return PrecommitNotInstalled()
-
-    # Launch install-hooks in the background to eagerly pre-install hook
-    # environments (especially the ansible language:python venv). Without this,
-    # the first commit pays the cost of creating the venv and downloading ansible.
-    # pre-commit uses flock on ~/.cache/pre-commit/.lock, so this is safe to run
-    # concurrently with a hook-triggered run.
-    log_dir = Path.home() / ".cache" / "claude-hooks"
-    log_dir.mkdir(parents=True, exist_ok=True)
-    log_fh = (log_dir / "pre-commit-install-hooks.log").open("w")
-    proc = subprocess.Popen(
-        [precommit, "install-hooks"], cwd=project_dir, stdout=log_fh, stderr=subprocess.STDOUT, start_new_session=True
-    )
-    logger.info(
-        "Started background pre-commit install-hooks (pid %d), log: %s",
-        proc.pid,
-        log_dir / "pre-commit-install-hooks.log",
-    )
-    return PrecommitInstallingHooks(pid=proc.pid)
 
 
 class LogCollector(logging.handlers.MemoryHandler):
@@ -476,7 +395,7 @@ async def run_web_mode(hook_input: HookInput, settings: HookSettings) -> None:
     results = await asyncio.gather(
         setup_proxy_with_supervisor(),
         setup_podman_with_supervisor(),
-        run_in_thread(install_git_precommit_hook, project_dir),
+        run_in_thread(precommit_setup.install_precommit, project_dir),
         run_in_thread(nix_setup.install_nix, settings),
         run_in_thread(install_bazelisk_wrapper),
         run_in_thread(setup_buildbuddy),
@@ -485,7 +404,7 @@ async def run_web_mode(hook_input: HookInput, settings: HookSettings) -> None:
     # Unpack with explicit type annotations for mypy
     proxy_result: proxy_setup.ProxySetup | BaseException = results[0]
     podman_result: podman_service.PodmanSetup | BaseException = results[1]
-    precommit_result: PrecommitSetup | BaseException = results[2]
+    precommit_result: precommit_setup.PrecommitSetup | BaseException = results[2]
     nix_result: Path | None | BaseException = results[3]
     bazelisk_result: bazelisk_setup.BazeliskSetup | BaseException = results[4]
     buildbuddy_result: str | None | BaseException = results[5]
